Soft delete topbar (#104)
* soft delete feature

* soft delete

* rm dev tools

* rm dev tools

* revert package lock

* soft delete

* add topbar controls

* layout improvements

* delete all function

* only update non-deleted items

---------

+++ httpSrv.ts
@@ -68,7 +68,7 @@ export class HttpServer {
 
 				console.error("DB get addresses fail")
 				console.error(err)
-
+				res.status(500).send('Failed to get addresses');
 			}
 
 		})
@@ -101,6 +101,7 @@ export class HttpServer {
 			} catch (err) {
 				console.error("DB get addresses fail")
 				console.error(err)
+				res.status(500).send('Failed to get address');
 			}
 		})
 
@@ -116,17 +117,17 @@ export class HttpServer {
 
 					res.status(200).send("exist");
 
+				} else {
+					this.db.prepare("INSERT INTO address (addr) VALUES (?)").run(address);
+					res.status(200).send("done");
 				}
-
-				this.db.prepare("INSERT INTO address (addr) VALUES (?)").run(address);
-				res.status(200).send("done");
 
 			} catch (err) {
 
 				console.error("DB add addresses fail")
 				console.error(err)
+				res.status(500).json({ error: "Failed to add address" });
 			}
-
 		})
 
 		app.post('/deleteAddress', (req, res) => {
@@ -141,16 +142,21 @@ export class HttpServer {
 				res.status(200).send("done");
 
 			} catch (err) {
-
-				console.error("DB delete address fail")
+				console.error("DB delete address fail");
 				console.error(err)
+				res.status(500).send('Failed to delete address');
 			}
 
 		})
 
 		app.post('/mails', (req, res) => {
+			// Atomics.wait(new Int32Array(new SharedArrayBuffer(4)), 0, 0, 1110);
 
-			const json = req.body;
+			const json = req.body as {
+				cursorId: string,
+				addr: string,
+				deleted: boolean
+			};
 
 			try {
 				const perPage = Number(config.getConfig('MailCountPerPage'));
@@ -166,13 +172,17 @@ export class HttpServer {
 				};
 
 				const comparisonOperator = direction === 'lt' ? '<' : '>';
-				const whereClause = cursorId ? `AND Id ${comparisonOperator} @cursorId` : '';
+				const whereClause =
+					(json.deleted ? ` deleted = 1 ` : ' deleted <> 1 ') +
+					(cursorId ? ` AND Id ${comparisonOperator} @cursorId` : '') +
+					(json.addr ? ` AND recipient = @recipient` : '')
+
 				const sortOrder = direction === 'lt' ? 'DESC' : 'ASC';
 
 				const sql = `
 				  SELECT id, sender, subject, read, received 
 				  FROM mail 
-				  WHERE recipient = @recipient ${whereClause}
+				  WHERE ${whereClause}
 				  ORDER BY id ${sortOrder} 
 				  LIMIT @mailCount
 				`;
@@ -190,10 +200,9 @@ export class HttpServer {
 				});
 
 			} catch (err) {
-
-				console.error("DB get mails fail")
-				console.error(err)
-
+				console.error("DB get mails fail");
+				console.error(err);
+				res.status(500).json({ error: "Failed to get mails" });
 			}
 
 		});
@@ -211,7 +220,7 @@ export class HttpServer {
 
 				console.error("DB get mail data fail")
 				console.error(err)
-
+				res.status(500).json({ error: "Failed to delete mails" });
 			}
 
 		});
@@ -222,16 +231,47 @@ export class HttpServer {
 
 			try {
 
-				this.db.prepare("DELETE FROM mail WHERE id = ?").run(json.id);
+				const mail = this.db.prepare("SELECT deleted FROM mail WHERE id = ?").get(json.id);
+				const isDeleted = mail as unknown as { deleted: boolean };
+
+				if (isDeleted.deleted) {
+					this.db.prepare("DELETE FROM mail WHERE id = ?").run(json.id);
+				}
+				else {
+					this.db.prepare("UPDATE mail SET deleted = 1 WHERE id = ?").run(json.id);
+				}
 				res.status(200).send();
 
 			} catch (err) {
 
 				console.error("DB delete mail fail")
 				console.error(err)
-
+				res.status(500).json({ error: "Failed to delete mail" });
 			}
 
+		})
+
+		app.post('/deleteMails', (req, res) => {
+			const json = req.body;
+			try {
+				this.db.prepare("UPDATE mail SET deleted = 1 WHERE recipient = ? and deleted = 0").run(json.address);
+				res.status(200).send();
+			} catch (err) {
+				console.error("DB delete mails fail")
+				console.error(err)
+				res.status(500).json({ error: "Failed to delete mails" });
+			}
+		})
+
+		app.post('/emptyDeletedMails', (req, res) => {
+			try {
+				this.db.prepare("DELETE FROM mail WHERE deleted = 1").run();
+				res.status(200).send();
+			} catch (err) {
+				console.error("DB empty deleted mails fail")
+				console.error(err)
+				res.status(500).json({ error: "Failed to empty deleted mails" });
+			}
 		})
 
 		app.post('/readMail', (req, res) => {
@@ -244,19 +284,20 @@ export class HttpServer {
 			} catch (err) {
 				console.error("DB update mail fail")
 				console.error(err)
+				res.status(500).json({ error: "Failed to update mail as read" });
 			}
 		})
 
 		app.post('/readAllMail', (req, res) => {
 
 			const json = req.body;
-
 			try {
 				this.db.prepare("UPDATE mail SET read = 1 where recipient = ? and read = 0").run(json.address);
 				res.status(200).send();
 			} catch (err) {
 				console.error("DB read all mail fail")
 				console.error(err)
+				res.status(500).json({ error: "Failed to mark all mails as read" });
 			}
 		})
 
@@ -266,13 +307,14 @@ export class HttpServer {
 				const unread = this.db.prepare(`
 					SELECT recipient, count(*) as unread
 					from mail
-					where read = 0
+					where read = 0 and deleted = 0
 					group by recipient
 					`).all();
 				res.json(unread);
 			} catch (err) {
 				console.error("unread counts select fail");
 				console.error(err)
+				res.status(500).json({ error: "Failed to get unread counts" });
 			}
 		})
